REFINE: Restructure UI packages and add BottomNavItem model
