+++ app/src/main/java/com/example/unimarket/MainActivity.kt
@@ -14,11 +14,11 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.example.unimarket.ui.data.PreferencesManager
-import com.example.unimarket.ui.login.LoginScreen
-import com.example.unimarket.ui.main.MainScreen
-import com.example.unimarket.ui.onboarding.OnboardingScreen
-import com.example.unimarket.ui.onboarding.PersonalizationScreen
-import com.example.unimarket.ui.register.RegisterScreen
+import com.example.unimarket.ui.views.LoginScreen
+import com.example.unimarket.ui.views.MainScreen
+import com.example.unimarket.ui.views.OnboardingScreen
+import com.example.unimarket.ui.views.PersonalizationScreen
+import com.example.unimarket.ui.views.RegisterScreen
 import com.example.unimarket.ui.theme.UniMarketTheme
 import com.google.firebase.analytics.FirebaseAnalytics
 import com.google.firebase.auth.FirebaseAuth
