+++ app/src/main/java/com/example/unimarket/ui/models/Major.kt
@@ -1,4 +1,4 @@
-package com.example.unimarket.ui.explore
+package com.example.unimarket.ui.models
 
 data class Major(
     val id: String = "",
