+++ app/src/main/java/com/example/unimarket/ui/viewModels/ExploreViewModel.kt
@@ -1,8 +1,10 @@
-package com.example.unimarket.ui.explore
+package com.example.unimarket.ui.viewModels
 
 import android.os.Bundle
 import androidx.lifecycle.ViewModel
 import com.example.unimarket.ui.data.FirebaseFirestoreSingleton
+import com.example.unimarket.ui.models.Product
+import com.example.unimarket.ui.models.User
 import com.google.firebase.analytics.FirebaseAnalytics
 import com.google.firebase.analytics.ktx.analytics
 import com.google.firebase.auth.FirebaseAuth
