+++ app/src/main/java/com/example/unimarket/ui/views/ExploreScreen.kt
@@ -1,4 +1,4 @@
-package com.example.unimarket.ui.explore
+package com.example.unimarket.ui.views
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
@@ -22,6 +22,7 @@ import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
+import androidx.compose.material3.SnackbarDuration
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Text
@@ -34,11 +35,14 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import coil.compose.rememberAsyncImagePainter
-import com.example.unimarket.ui.sensor.ShakeDetector
+import com.example.unimarket.ui.viewModels.ExploreViewModel
+import com.example.unimarket.ui.models.Product
+import com.example.unimarket.ui.viewModels.ShakeDetector
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 
@@ -61,7 +65,7 @@ fun ExploreScreen(
             delay(10000) // Wait 10 seconds
             snackbarHostState.showSnackbar(
                 message = "Tip: Shake your phone to refresh products.",
-                duration = androidx.compose.material3.SnackbarDuration.Short
+                duration = SnackbarDuration.Short
             )
             tipShown.value = true
         }
@@ -73,7 +77,7 @@ fun ExploreScreen(
         coroutineScope.launch {
             snackbarHostState.showSnackbar(
                 message = "Products refreshed!",
-                duration = androidx.compose.material3.SnackbarDuration.Short
+                duration = SnackbarDuration.Short
             )
         }
     }
@@ -181,7 +185,7 @@ fun ProductCard(product: Product) {
         Column(modifier = Modifier.padding(16.dp)) {
             val painter = rememberAsyncImagePainter(
                 model = product.imageUrls.firstOrNull(),
-                contentScale = androidx.compose.ui.layout.ContentScale.Crop
+                contentScale = ContentScale.Crop
             )
             Image(
                 painter = painter,
@@ -189,7 +193,7 @@ fun ProductCard(product: Product) {
                 modifier = Modifier
                     .fillMaxSize() // Fill the size of the Card
                     .height(180.dp),
-                contentScale = androidx.compose.ui.layout.ContentScale.Crop
+                contentScale = ContentScale.Crop
             )
             Spacer(modifier = Modifier.height(8.dp))
             Text(text = product.title, style = MaterialTheme.typography.titleMedium)
