+++ app/src/main/java/com/example/unimarket/ui/views/LoginScreen.kt
@@ -1,4 +1,4 @@
-package com.example.unimarket.ui.login
+package com.example.unimarket.ui.views
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
@@ -29,6 +29,7 @@ import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.example.unimarket.R
+import com.example.unimarket.ui.viewModels.LoginViewModel
 
 @Composable
 fun LoginScreen(
